Add files via upload
Object:
Predict the percentage of a student based on the no. of study hours. This is a simple linear regression task as it involves just 2 variables. You can use R, Python, SAS Enterprise Miner or any other tool. What will be the predicted score if a student studies for 9.25 hrs/ day?
I achieve a great predicted Score = 93.69173248737538